requirements.txt
streamlit==1.28.0
opencv-python-headless==4.8.1.78
Pillow==10.0.1
numpy==1.24.3